Update home screen to accept vm state data
