+++ feature/home/src/main/java/com/epmedu/animeal/home/presentation/HomeScreenEvent.kt
@@ -1,0 +1,6 @@
+package com.epmedu.animeal.home.presentation
+
+internal sealed interface HomeScreenEvent {
+    data class FeedSpotSelected(val id: Int = -1) : HomeScreenEvent
+    data class FeedSpotFavouriteChange(val id: Int = -1, val isFavourite: Boolean) : HomeScreenEvent
+}
