+++ feature/home/src/main/java/com/epmedu/animeal/home/presentation/viewmodel/HomeEvent.kt
@@ -1,0 +1,5 @@
+package com.epmedu.animeal.home.presentation.viewmodel
+
+internal sealed interface HomeEvent {
+    object ShowCurrentFeedSpotInformation : HomeEvent
+}
