+++ feature/more/src/main/java/com/epmedu/animeal/more/root/MoreScreenUi.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.epmedu.animeal.foundation.theme.AnimealTheme
+import com.epmedu.animeal.foundation.theme.bottomBarPadding
 import com.epmedu.animeal.foundation.topbar.TopBar
 import com.epmedu.animeal.more.root.ui.MoreOption
 import com.epmedu.animeal.more.screens
@@ -24,7 +25,8 @@ internal fun MoreScreenUi(
 ) {
     Scaffold(
         modifier = Modifier
-            .fillMaxSize(),
+            .fillMaxSize()
+            .bottomBarPadding(),
         topBar = {
             TopBar(
                 title = stringResource(id = R.string.more),
