+++ library/foundation/src/main/java/com/epmedu/animeal/foundation/animation/VerticalSlideAnimatedVisibility.kt
@@ -1,0 +1,20 @@
+package com.epmedu.animeal.foundation.animation
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.AnimatedVisibilityScope
+import androidx.compose.animation.slideInVertically
+import androidx.compose.animation.slideOutVertically
+import androidx.compose.runtime.Composable
+
+@Composable
+fun VerticalSlideAnimatedVisibility(
+    visible: Boolean,
+    content: @Composable AnimatedVisibilityScope.() -> Unit
+) {
+    AnimatedVisibility(
+        visible = visible,
+        enter = slideInVertically { it },
+        exit = slideOutVertically { it },
+        content = content
+    )
+}
