+++ library/foundation/src/main/java/com/epmedu/animeal/foundation/theme/Modifier.kt
@@ -1,0 +1,6 @@
+package com.epmedu.animeal.foundation.theme
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.ui.Modifier
+
+fun Modifier.bottomBarPadding(): Modifier = this.padding(bottom = bottomBarHeight)
