+++ library/foundation/src/main/java/com/epmedu/animeal/foundation/theme/Size.kt
@@ -1,0 +1,5 @@
+package com.epmedu.animeal.foundation.theme
+
+import androidx.compose.ui.unit.dp
+
+val bottomBarHeight = 56.dp
